Move pkg/predicate closer to where it is used.
