+++ pkg/ansible/predicate/predicate.go
